planner: switch to APS for path planning
RRT* is sometimes unable to find a solution within ~6s in cases where a
large amount of shuffling is required to get around obstacles.

Switch to using APS with LBKPIECE, which combines the "guaranteed"
(or nearly so) path finding of LBKPIECE with path shortening to approach
RRT* behaviour with a greater chance of finding complete paths.